+++ mzmine-community/src/main/java/io/github/mzmine/modules/dataprocessing/featdet_baselinecorrection/BaselineCorrectionParameters.java
@@ -39,6 +39,7 @@ import java.util.Collection;
 import java.util.stream.Collectors;
 import javafx.application.Platform;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class BaselineCorrectionParameters extends SimpleParameterSet {
 
@@ -91,5 +92,19 @@ public class BaselineCorrectionParameters extends SimpleParameterSet {
         BaselineCorrectionPreview::new);
     dialog.showAndWait();
     return dialog.getExitCode();
+  }
+
+  @Override
+  public int getVersion() {
+    return 2;
+  }
+
+  @Override
+  public @Nullable String getVersionMessage(int version) {
+    return switch (version) {
+      case 2 ->
+          "Baseline correction parameters were updated to use a relative percentage of sampled data points instead of an absolute number.";
+      default -> null;
+    };
   }
 }
